+++ cmd/exporters.go
@@ -62,15 +62,18 @@ var exportersListCmd = &cobra.Command{
 	Short: "List all available exporters",
 	Run: func(cmd *cobra.Command, args []string) {
 		for _, exporter := range ortfodb.BuiltinExporters() {
-			printExporter(exporter)
+			fmt.Print(showExporter(exporter))
 		}
 	},
 }
 
-var exporterHelpCmd = &cobra.Command{
-	Use:   "help <name>",
-	Short: "Get help for a specific exporter",
-	Args:  cobra.ExactArgs(1),
+var exampleExporter = ortfodb.BuiltinNativeExporters[1]
+
+var exporterDocCmd = &cobra.Command{
+	Use:     "doc <name>",
+	Short:   "Get help for a specific exporter",
+	Example: strings.Join([]string{"$ ortfodb exporters help " + exampleExporter.Name(), "", showExporter(exampleExporter), howToAdd(exampleExporter, &pflag.FlagSet{})}, "\n"),
+	Args:    cobra.ExactArgs(1),
 	ValidArgsFunction: func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
 		exporters := make([]string, 0, len(ortfodb.BuiltinExporters()))
 		for _, exporter := range ortfodb.BuiltinExporters() {
@@ -81,8 +84,8 @@ var exporterHelpCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		for _, exporter := range ortfodb.BuiltinExporters() {
 			if exporter.Name() == args[0] {
-				printExporter(exporter)
-				howToAdd(exporter, cmd.Flags())
+				fmt.Print(showExporter(exporter))
+				fmt.Print(howToAdd(exporter, cmd.Flags()))
 				return
 			}
 		}
@@ -92,7 +95,7 @@ var exporterHelpCmd = &cobra.Command{
 func init() {
 	exportersCmd.AddCommand(exportersInitCmd)
 	exportersCmd.AddCommand(exportersListCmd)
-	exportersCmd.AddCommand(exporterHelpCmd)
+	exportersCmd.AddCommand(exporterDocCmd)
 	rootCmd.AddCommand(exportersCmd)
 }
 
@@ -117,58 +120,62 @@ func exporterDetails(exporter ortfodb.Exporter) (name, description string, requi
 	}
 }
 
-func howToAdd(exporter ortfodb.Exporter, flags *pflag.FlagSet) {
+func howToAdd(exporter ortfodb.Exporter, flags *pflag.FlagSet) string {
+	output := ""
 	name, _, _, options := exporterDetails(exporter)
 	if len(options) == 0 {
-		return
+		return output
 	}
 	configFilename, _ := flags.GetString("config")
 	if configFilename == "" {
 		configFilename = "your ortfodb config file"
 	}
 
-	fmt.Printf(colorstring.Color("To add [bold]%s[reset] to your project, add the following to [cyan]%s[reset]:\n\n"), name, configFilename)
+	output += fmt.Sprintf(colorstring.Color("To add [bold]%s[reset] to your project, add the following to [cyan]%s[reset]:\n\n"), name, configFilename)
 
-	fmt.Printf(colorstring.Color("  [bold][dim][red]exporters:\n[reset]    [bold][red]%s:[reset] [dim]# <- add this alongside your potential other exporters\n"), name)
+	output += fmt.Sprintf(colorstring.Color("  [bold][dim][red]exporters:\n[reset]    [bold][red]%s:[reset] [dim]# <- add this alongside your potential other exporters\n"), name)
 	for key, defaultValue := range options {
 		renderedDefaultValueBytes, _ := json.Marshal(defaultValue)
 		renderedDefaultValue := string(renderedDefaultValueBytes)
 		if renderedDefaultValue == "null" {
 			renderedDefaultValue = ""
 		}
-		fmt.Printf(colorstring.Color("      [bold][red]%s:[reset] [green]%s[reset]\n"), key, renderedDefaultValue)
+		output += fmt.Sprintf(colorstring.Color("      [bold][red]%s:[reset] [green]%s[reset]\n"), key, renderedDefaultValue)
 	}
 
-	fmt.Println("\nFeel free to change these configuration values. Check out the exporter's documentation to learn more about what they do.")
+	output += ("\nFeel free to change these configuration values. Check out the exporter's documentation to learn more about what they do.\n")
+	return output
 }
 
-func printExporter(exporter ortfodb.Exporter) {
+func showExporter(exporter ortfodb.Exporter) string {
+	output := ""
 	name, description, requires, options := exporterDetails(exporter)
 	wrappedDescription := wrap(12, terminalWidth(20, 100), description)
-	fmt.Println(colorstring.Color(fmt.Sprintf("[bold][blue]%-10s[reset]  %s", name, wrappedDescription)))
+	output += (colorstring.Color(fmt.Sprintf("[bold][blue]%-10s[reset]  %s\n", name, wrappedDescription)))
 	hasDetails := false
 	descriptionIsMultiline := strings.Contains(wrappedDescription, "\n")
 	if len(requires) > 0 {
 		if descriptionIsMultiline {
-			fmt.Println()
+			output += "\n"
 		}
-		fmt.Println(colorstring.Color(fmt.Sprintf("%12s[bold][yellow]Requires[reset]: %s", "", strings.Join(requires, ", "))))
+		output += colorstring.Color(fmt.Sprintf("%12s[bold][yellow]Requires[reset]: %s\n", "", strings.Join(requires, ", ")))
 		hasDetails = true
 	}
 	if len(keys(options)) > 0 {
 		if descriptionIsMultiline && !hasDetails {
 			fmt.Println()
 		}
-		fmt.Printf(colorstring.Color("%12s[bold][cyan]Options[reset]:\n"), "")
+		output += fmt.Sprintf(colorstring.Color("%12s[bold][cyan]Options[reset]:\n"), "")
 		optionKeys := keys(options)
 		sort.Strings(optionKeys)
 		for _, key := range optionKeys {
 			// TODO add a way to add descriptions to options
-			fmt.Println(colorstring.Color(fmt.Sprintf("%12s[bold][dim]•[reset] [blue]%s[reset] %s", "", key, "")))
+			output += (colorstring.Color(fmt.Sprintf("%12s[bold][dim]•[reset] [blue]%s[reset] %s\n", "", key, "")))
 		}
 		hasDetails = true
 	}
 	if hasDetails {
-		fmt.Println()
+		output += "\n"
 	}
+	return output
 }